incorporated advanced game into lobby
